Split comtrollers into app and ephemeral
