+++ lib/controllers/ephemeral/splash_controller.dart
